sphere

+++ bagel_render_3d/src/options.rs
@@ -1,10 +1,10 @@
 pub struct ConsoleOptions {
-    pub width: i64,
-    pub height: i64,
+    pub width: u16,
+    pub height: u16,
 }
 
 impl ConsoleOptions {
-    pub fn new(width: i64, height: i64) -> Self {
+    pub fn new(width: u16, height: u16) -> Self {
         ConsoleOptions { width, height }
     }
 }
